do not return any js type

+++ wasm/consts.rs
@@ -1,7 +1,7 @@
 use crate::types::sqrt_price::get_max_tick;
 use js_sys::BigInt;
 use wasm_bindgen::prelude::*;
-use wasm_wrapper::wasm_wrapper;
+
 pub const MAX_TICK: i32 = 221_818;
 pub const MIN_TICK: i32 = -MAX_TICK;
 
@@ -23,22 +23,22 @@ pub const MAX_POOL_KEYS_RETURNED: u16 = 220;
 pub const MAX_POOL_PAIRS_RETURNED: usize =
     MAX_RESULT_SIZE / (128 + 128 + 32 + 128 + 128 + 128 + 128 + 64 + 64 + 32 + 64 + 16);
 
-#[wasm_wrapper]
-pub fn get_global_max_sqrt_price() -> u128 {
-    MAX_SQRT_PRICE
+#[wasm_bindgen]
+pub fn get_global_max_sqrt_price() -> BigInt {
+    BigInt::from(MAX_SQRT_PRICE)
 }
 
-#[wasm_wrapper]
-pub fn get_global_min_sqrt_price() -> u128 {
-    MIN_SQRT_PRICE
+#[wasm_bindgen]
+pub fn get_global_min_sqrt_price() -> BigInt {
+    BigInt::from(MIN_SQRT_PRICE)
 }
 
-#[wasm_wrapper]
+#[wasm_bindgen]
 pub fn get_tick_search_range() -> i32 {
     TICK_SEARCH_RANGE
 }
 
-#[wasm_wrapper]
+#[wasm_bindgen]
 pub fn get_max_chunk(tick_spacing: u16) -> u16 {
     let max_tick = get_max_tick(tick_spacing);
     let max_bitmap_index = (max_tick + MAX_TICK) / tick_spacing as i32;
@@ -46,32 +46,32 @@ pub fn get_max_chunk(tick_spacing: u16) -> u16 {
     max_chunk_index as u16
 }
 
-#[wasm_wrapper]
+#[wasm_bindgen]
 pub fn get_chunk_size() -> i32 {
     CHUNK_SIZE
 }
 
-#[wasm_wrapper]
+#[wasm_bindgen]
 pub fn get_max_tick_cross() -> i32 {
     MAX_TICK_CROSS
 }
 
-#[wasm_wrapper]
-pub fn get_max_tickmap_query_size() -> u64 {
-    MAX_TICKMAP_QUERY_SIZE as u64
+#[wasm_bindgen]
+pub fn get_max_tickmap_query_size() -> u32 {
+    MAX_TICKMAP_QUERY_SIZE as u32
 }
 
-#[wasm_wrapper]
-pub fn get_liquidity_ticks_limit() -> u64 {
-    LIQUIDITY_TICK_LIMIT as u64
+#[wasm_bindgen]
+pub fn get_liquidity_ticks_limit() -> u32 {
+    LIQUIDITY_TICK_LIMIT as u32
 }
 
-#[wasm_wrapper]
+#[wasm_bindgen]
 pub fn get_max_pool_keys_returned() -> u16 {
     MAX_POOL_KEYS_RETURNED
 }
 
-#[wasm_wrapper]
-pub fn get_max_pool_pairs_returned() -> u64 {
-    MAX_POOL_PAIRS_RETURNED as u64
+#[wasm_bindgen]
+pub fn get_max_pool_pairs_returned() -> u32 {
+    MAX_POOL_PAIRS_RETURNED as u32
 }
